refactor(domain): domain 파일 이동
- Reservation.java
- Theme.java
- TimeTable.java
