+++ src/main/java/nextstep/domain/Reservation.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
